+++ Client/src/components/TabModelling/TabModelling.js
@@ -5,7 +5,7 @@ import TabModellingInputs from './TabModellingInputs';
 import TabModellingAdvanced from './TabModellingAdvanced';
 import TabModellingRunMain from './TabModellingRunMain';
 import TabModellingRunBootstrap from './TabModellingRunBootstrap';
-import TabModellingTables from './TabModellingTables';
+import TabModellingOutputs from './TabModellingOutputs';
 
 const TabModelling = (props) => {
   const { appMgr } = props;
@@ -19,7 +19,7 @@ const TabModelling = (props) => {
       {activeSubPageId === 2 && <TabModellingAdvanced {...props} />}
       {activeSubPageId === 3 && <TabModellingRunMain {...props} />}
       {activeSubPageId === 4 && <TabModellingRunBootstrap {...props} />}
-      {activeSubPageId === 5 && <TabModellingTables {...props} />}
+      {activeSubPageId === 5 && <TabModellingOutputs {...props} />}
     </React.Fragment>
   );
 };